+++ forward/src/assets/app/settings.js
@@ -117,6 +117,7 @@ export class AppSettings {
                 volumeSizeDefault: new NumberSetting(40, { 'choises': [1, 10, 20, 30, 40, 50] }),
                 volumeSizeMin: new NumberSetting(40, { 'choises': [1, 10, 20, 30, 40, 50] }),
                 imageUploadBlockSize: new NumberSetting(10, { 'choises': [1, 5, 10, 20, 40, 80, 160] }),
+                supportResourceAction: new BooleanSetting(false),
             }
         )
     }